+++ spec/spec_helper.rb
@@ -1,3 +1,3 @@
 $:.unshift(File.expand_path("../lib", File.dirname(__FILE__)))
-require "octopress-pygments"
+require "octopress-code-highlighter"